+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Orange and deep orange palettes and swatches - [#42](https://github.com/dartoos-dev/eo_color/issues/42).
+  - **Palettes:** Orange, OrangeAccent; DeepOrange, DeepOrangeAccent.
+  - **Swatches:** Oranges, OrangeAccents; DeepOranges, DeepOrangeAccents.
 
 ## [0.0.6] - 2021-05-19